types mostly done, still a potential bug with negative costs

+++ bellmanford_c/v2/include/graph.h
@@ -31,7 +31,7 @@ typedef struct ford{
 }ford_t;
 
 typedef struct mcost{
-    int32_t cost;
+    int64_t cost;
     uint32_t node;
 } mcost_t;
